requirements.txt
pytest==7.3.1
Appium-Python-Client==2.10.1
environs==9.5.0
Faker==18.10.1
